updated blur (with temporary solution) in fit_hysteresis
